Remove Q_OBJECT macro from interface header
The macro is removed from the interface header, since QT generates the
moc files from the files participating in the project. If we have the
macro in the interface header, we will need to add the header into the
lib sources list as well, therefore the header will lose its purpose.